\name{arthritis}
\alias{arthritis}
\docType{data}
\title{The rheumatoid arthritis data}

\description{ Data obtained from a meta-analysis that aimed to determine whether anti-cyclic citrullinated peptide (anti-CCP) antibody  identifies more accurately patients with rheumatoid arthritis than rheumatoid factor (RF) does. We include \eqn{N=22}  studies that assessed both RF and anti-CCP2 antibody for diagnosing rheumatoid arthritis.
}
\format{
  A data frame with 22 observations on the following 8 variables.
  \describe{
    \item{TP1}{the number of true positives for RF}
\item{FN1}{the number of false negatives for RF}
\item{FP1}{the number of false positives for RF}
\item{TN1}{the number of true negatives for RF}
 \item{TP2}{the number of true positives for anti-CCP2}
\item{FN2}{the number of false negatives for anti-CCP2}
\item{FP2}{the number of false positives for anti-CCP2}
\item{TN2}{the number of true negatives for anti-CCP2}
}}
\references{
Nishimura, K., Sugiyama, D., Kogata, Y., et al. (2007) Meta-analysis: Diagnostic accuracy of anti-cyclic citrullinated peptide antibody and rheumatoid factor for rheumatoid arthritis. \emph{Annals of Internal Medicine}, \bold{146}(11), 
797--808.

Dimou, N.L., Adam, M. and Bagos, P.G. (2016) A multivariate method for meta-analysis and comparison of diagnostic
tests. \emph{Statistics in Medicine}, \bold{35}(20), 3509--3523.

}

\keyword{datasets}

