\name{coronary}
\alias{coronary}
\docType{data}
\title{The coronary CT angiography data}

\description{Data on 26 studies from a systematic review for diagnostic accuracy studies of coronary computed tomography angiography for the detection of coronary artery disease.  
}
\usage{data(coronary)}
\format{
  A data frame with 26 observations on the following 6 variables.
  \describe{
    \item{TP}{the number of true positives}
\item{FN}{the number of false negatives}
\item{FP}{the number of false positives}
\item{TN}{the number of true negatives}
\item{NEP}{the number of non-evaluable positives}
\item{NEN}{the number of non-evaluable negatives}
}
}

\references{
Schuetz, G. M., Schlattmann, P., and Dewey, M. (2012). Use of 3x2 tables with an intention to diagnose approach to assess clinical performance of diagnostic tests: Meta-analytical evaluation of coronary CT angiography studies. \emph{BMJ} (Online), \bold{345}:e6717.


}

\keyword{datasets}


