\name{Down}
\alias{Down}
\docType{data}
\title{The down syndrome data}

\description{Data on \eqn{N=11} studies from the systematic review that examined the screening accuracy of shortened humerus and shortened femur of the fetus markers (two out of seven ultrasonographic markers or their combination in detecting Down syndrome in Smith-Bindman et al., 2001).
}
\format{
  A data frame with 11 observations on the following 8 variables.
  \describe{
    \item{down_n_00}{the number of the test results in the diseased where the shortened humerus outcome is negative and the shortened femur outcome is negative}
\item{down_n_01}{the number of the test results in the diseased where the shortened humerus outcome is negative and the shortened femur outcome is positive}
\item{down_n_10}{the number of the test results in the diseased where the shortened humerus outcome is positive and the shortened femur outcome is negative}
\item{down_n_11}{the number of the test results in the diseased where the shortened humerus outcome is positive and the shortened femur outcome is positive}
\item{nodown_n_00}{the number of the test results in the non-diseased where the shortened humerus outcome is negative and the shortened femur outcome is negative}
\item{nodown_n_01}{the number of the test results in the non-diseased where the shortened humerus outcome is negative and the shortened femur outcome is positive}
\item{nodown_n_10}{the number of the test results in the non-diseased where the shortened humerus outcome is positive and the shortened femur outcome is negative}
\item{nodown_n_11}{the number of the test results in the non-diseased where the shortened humerus outcome is positive and the shortened femur outcome is positive}
}}
\references{
Smith-Bindman R, Hosmer W, Feldstein V et al. Second-trimester ultrasound to detect fetuses with down syndrome: A meta-analysis (2001). \emph{Journal of the American Medical Association}, \bold{285}(8): 1044-1055.
}

\keyword{datasets}

