\name{mgrid}
\alias{mgrid}
\alias{mgrid15}
\alias{mgrid30}
\alias{mgrid50}
\docType{data}
\title{A list containing four-dimensional arrays}

\description{A list containing four-dimensional arrays.  Replicates of the quadrature points  that produce a 4-dimensional full grid.}


\examples{
data(mgrid15)
dim(mgrid15$x)
dim(mgrid15$y)
dim(mgrid15$z)
dim(mgrid15$w)

data(mgrid30)
dim(mgrid30$x)
dim(mgrid30$y)
dim(mgrid30$z)
dim(mgrid30$w)

data(mgrid50)
dim(mgrid50$x)
dim(mgrid50$y)
dim(mgrid50$z)
dim(mgrid50$w)
}
\keyword{datasets}

