\name{mgrid5d}
\alias{mgrid5d}
\alias{mgrid5d15}
\alias{mgrid5d30}
\docType{data}
\title{A list containing five-dimensional arrays}

\description{A list containing five-dimensional arrays.  Replicates of the quadrature points  that produce a 5-dimensional full grid.}


\examples{
data(mgrid5d15)
dim(mgrid5d15$x1)
dim(mgrid5d15$x2)
dim(mgrid5d15$x3)
dim(mgrid5d15$x4)
dim(mgrid5d15$x5)

data(mgrid5d30)
dim(mgrid5d30$x1)
dim(mgrid5d30$x2)
dim(mgrid5d30$x3)
dim(mgrid5d30$x4)
dim(mgrid5d30$x5)


}
\keyword{datasets}

