\name{mgrid6d}
\alias{mgrid6d}
\docType{data}
\title{A list containing six-dimensional arrays}

\description{A list containing six-dimensional arrays.  Replicates of the quadrature points  that produce an 6-dimensional full grid.}


\examples{
data(mgrid6d)
dim(mgrid$x1)
dim(mgrid$x2)
dim(mgrid$x3)
dim(mgrid$x4)
dim(mgrid$x5)
dim(mgrid$x6)


}
\keyword{datasets}

