\name{MK2016}
\alias{MK2016}
\docType{data}
\title{The coronary CT angiography data in Menke and Kowalski (2016).}

\description{Data on 30 studies from a systematic review for diagnostic accuracy studies of coronary computed tomography angiography for the detection of coronary artery disease.  
}
\usage{data(MK2016)}
\format{
  A data frame with 30 observations on the following 6 variables.
  \describe{
    \item{TP}{the number of true positives}
\item{FN}{the number of false negatives}
\item{FP}{the number of false positives}
\item{TN}{the number of true negatives}
\item{NEP}{the number of non-evaluable positives}
\item{NEN}{the number of non-evaluable negatives}
}
}

\references{
Menke, J. and Kowalski, J. (2016). Diagnostic accuracy and utility of coronary ct angiography with consideration of unevaluable results: A systematic review and multivariate bayesian random-effects meta-analysis with intention to diagnose. \emph{European Radiology}, \bold{26}(2):451--458.
}

\keyword{datasets}


