\name{MRI}
\alias{MRI}
\docType{data}
\title{The magnetic resonance imaging data}

\description{Data on 10 studies of magnetic resonance imaging (MRI) for the diagnosis of lymph node metastasis in women with cervical cancer, the last imaging technique in the meta-analysis in Scheidler et al. (1997). Diagnosis of metastatic disease by MRI relies on nodal enlargement. 
}
\usage{data(MRI)}
\format{
  A data frame with 10 observations on the following 4 variables.
  \describe{
    \item{TP}{the number of true positives}
\item{FN}{the number of false negatives}
\item{FP}{the number of false positives}
\item{TN}{the number of true negatives}
}}
\references{Scheidler, J., H. Hricak, K. K. Yu, L. Subak, and M. R. Segal (1997) Radiological evaluation of lymph node metastases in patients with cervical cancer: A meta-analysis.  \emph{Journal of the American Medical Association} 278: 1096--1101.
}

\keyword{datasets}
