\name{OGT}
\alias{OGT}
\docType{data}
\title{The orale glucose tolerance data}

\description{Data on 10 studies of  the oral glucose tolerance test for the diagnosis of diabetes mellitus in patients during acute coronary syndrome hospitalization in Ye et al. (2012). 
}
\usage{data(OGT)}
\format{
  A data frame with 10 observations on the following 4 variables.
  \describe{
    \item{TP}{the number of true positives}
\item{FN}{the number of false negatives}
\item{FP}{the number of false positives}
\item{TN}{the number of true negatives}
}}
\references{
Ye, Y., Xie, H., Zhao, X., Zhang, S. (2012) The oral glucose tolerance test for the diagnosis of diabetes mellitus in patients during acute coronary syndrome hospitalization: a meta-analysis of diagnostic test accuracy. \emph{Cardiovascular Diabetology}, \bold{11}(5):155.

}

\keyword{datasets}

