\name{Pap}
\alias{Pap}
\docType{data}
\title{The Pap test data}

\description{These data are comprised of \eqn{N=59} studies that published between January 1984 and March 1992. The diagnostic accuracy of the Pap test (i.e., index test) is evaluated by comparing with the histology test (i.e., reference test), which is not a perfect test (Fahey, et al., 1995).}
\format{
  A data frame with 59 observations on the following 4 variables.
  \describe{
\item{y11}{the number of the test results where the Pap test  outcome is positive and the  histology test outcome is positive}
\item{y10}{the number of the test results where the Pap test outcome is positive and the  histology test outcome is negative}
\item{y01}{the number of the test results where the Pap test outcome is negative and the  histology test outcome is positive}
\item{y00}{the number of the test results where the Pap test outcome is negative and the  histology test outcome is negative}
}}
\references{
Fahey, M. T., Irwig, L., and Macaskill, P. (1995). Meta-analysis of pap test accuracy. \emph{American
Journal of Epidemiology}, \bold{142}:680--689.
}

\keyword{datasets}

