\name{telomerase}
\alias{telomerase}
\docType{data}
\title{The telomerase data}

\description{In Glas et al. (2003) the telomerase marker for the diagnosis of bladder cancer is evaluated using 10 studies. The interest was to define if this non-invasive and cheap marker could replace the standard of cystoscopy or histopathology. 
}
\usage{data(telomerase)}
\format{
  A data frame with 10 observations on the following 4 variables.
  \describe{
    \item{TP}{the number of true positives}
\item{FN}{the number of false negatives}
\item{FP}{the number of false positives}
\item{TN}{the number of true negatives}
}}
\references{Glas A.S., Roos D., Deutekom M., Zwinderman A.H., Bossuyt P.M., Kurth K.H. (2003) Tumor markers in the diagnosis of primary bladder cancer. A systematic review. \emph{Journal of Urology} 169(6):1975--82.
}

\keyword{datasets}
