\name{betaDG}
\alias{betaDG}
\docType{data}
\title{The beta-D-Glucan-data}

\description{Data on 8 cohort studies inthemeta-analysis inKarageorgopouloset al.(2011). The interes there is to assess \eqn{beta}-D-Glucan as aserum or plasma marker for the presence of invasive fungal infections.}
\usage{data(betaDG)}
\format{
  A data frame with 8 observations on the following 4 variables.
  \describe{
    \item{TP}{the number of true positives}
\item{FN}{the number of false negatives}
\item{FP}{the number of false positives}
\item{TN}{the number of true negatives}
}}
\references{
Karageorgopoulos DE, Vouloumanou EK, Ntziora F,
Michalopoulos A, Rafailidis PI, Falagas ME. (2011) \eqn{beta}-D-Glucan assay for the diagnosis of invasive fungal infections: a meta-analysis. \emph{Clinical Infectious Diseases}, 52(6):750--770.
}

\keyword{datasets}

